requirements.txt
amqp==2.5.1
aniso8601==7.0.0
autopep8==1.4.4
babel==2.7.0
beautifulsoup4==4.7.1
billiard==3.6.1.0
bleach==2.1.4
boto3==1.9.247
botocore==1.12.247
braintree==3.49.0
cachetools==3.1.1
cairocffi==1.1.0
cairosvg==2.4.2
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.2
dataclasses==0.6; python_version < "3.7"
defusedxml==0.6.0
dj-database-url==0.5.0
dj-email-url==0.2.0
django==2.2.6
django-appconf==1.0.3
django-babel==0.6.2
django-bootstrap4==1.0.1
django-cache-url==3.0.0
django-countries==5.5
django-elasticsearch-dsl==0.5.1
django-filter==2.2.0
django-graphql-jwt==0.2.1
django-impersonate==1.4.1
django-js-asset==1.2.2
django-measurement==3.0.0
django-mptt==0.10.0
django-phonenumber-field==2.4.0
django-prices==2.1.0
django-prices-openexchangerates==1.0.1
django-prices-vatlayer==1.0.2
django-recaptcha==1.5.0
django-redis==4.10.0
django-render-block==0.6
django-silk==2.0.0
django-storages==1.7.2
django-templated-email==2.3.0
django-versatileimagefield==1.10
django-webpack-loader==0.6.0
docutils==0.15.2
draftjs-sanitizer==1.0.0
elasticsearch==6.3.1
elasticsearch-dsl==6.0.1
faker==2.0.2
freezegun==0.3.12
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.20.0
google-i18n-address==2.3.5
google-measurement-protocol==1.0.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
gprof2dot==2016.10.13
graphene==2.1.8
graphene-django==2.6.0
graphene-django-optimizer==0.6.1
graphene-federation==0.0.3
graphql-core==2.2.1
graphql-relay==2.0.0
html-to-draftjs==1.0.1
html5lib==1.0.1
idna==2.8
importlib-metadata==0.23
ipaddress==1.0.22
jinja2==2.10.3
jmespath==0.9.4
jsonfield==2.0.2
kombu==4.6.5
lxml==4.4.1
markdown==3.1.1
markupsafe==1.1.1
maxminddb==1.5.1
maxminddb-geolite2==2018.703
measurement==3.0.0
more-itertools==7.2.0
mpmath==1.1.0
oauthlib==3.1.0
phonenumberslite==8.10.20
pillow==5.4.1
prices==1.0.0
promise==2.2.1
protobuf==3.10.0
psycopg2-binary==2.8.3
purl==1.5
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pygments==2.4.2
pyjwt==1.7.1
pyphen==0.9.5
python-dateutil==2.8.0
python3-openid==3.1.0; python_version >= "3.0"
pytz==2019.3
razorpay==1.2.0
redis==3.3.10
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
rx==1.6.1
s3transfer==0.2.1
sentry-sdk==0.12.3
singledispatch==3.4.0.3
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.4
sqlparse==0.3.0
stripe==2.37.2
sympy==1.4
text-unidecode==1.3
tinycss2==1.0.2
tqdm==4.36.1
typing==3.7.4.1
urllib3==1.25.6
uwsgi==2.0.18; platform_system != "Windows"
vine==1.3.0
weasyprint==50
webencodings==0.5.1
zipp==0.6.0
nexmo
django-mathfilters